Add toml support
TOML files are read to dictionaries just like JSON or YAML, so to
support them as well.
TOML doesn't allow duplicate keys as per its standard, so no explicit
check is needed for that.

The `tomllib` package was added in Python 3.11.  For older versions, the
third-party package `tomli` can be used as replacement (`tomllib` is
actually based on `tomli`, so they should be pretty compatible).